### 2025-10-19 #267 @HzaCode
[SUBMISSION] OneCite — Universal citation & academic reference toolkit
Submitting Author: **(@HzaCode)**
All current maintainers: **(@HzaCode)**
Package Name: **OneCite**
One-Line Description of Package: **Universal citation & academic reference toolkit—multi-source lookup (DOI/PMID/arXiv/ISBN/GitHub/Zenodo/OpenAlex/Crossref/DataCite/Google Books), metadata completion, de-duplication, and multi-style output (BibTeX/APA/MLA); CLI + Python API.**
Repository Link: **[https://github.com/HzaCode/OneCite](https://github.com/HzaCode/OneCite)**
Version submitted: **0.0.12**
Archive: **TBD**
JOSS DOI: **N/A**
Version accepted: **TBD**
Date accepted (month/day/year): **TBD**

---

* [x] I agree to abide by [[pyOpenSci's Code of Conduct](https://www.pyopensci.org/handbook/CODE_OF_CONDUCT.html)][PyOpenSciCodeOfConduct] during the review process and in future interactions in spaces supported by pyOpenSci should it be accepted.
* [x] I have read and will commit to package maintenance after the review as per the [[pyOpenSci Policies Guidelines](https://www.pyopensci.org/software-peer-review/our-process/policies.html#after-acceptance-package-ownership-and-maintenance)][Commitment].

**OneCite** is a Python library and CLI that parses messy references, queries multiple scholarly databases, completes missing metadata, de-duplicates results, and exports standardized citations in **BibTeX/APA/MLA** through a templated pipeline (**clean → query → validate → format**). It supports identifiers and free-text (e.g., **DOI, PMID, arXiv, ISBN, GitHub, Zenodo**) and includes optional interactive disambiguation.

Please indicate which category or categories:

* [ ] Astropy: My package adheres to Astropy community standards
* [ ] Pangeo: My package adheres to the [[Pangeo standards listed in the pyOpenSci peer review guidebook](https://www.pyopensci.org/software-peer-review/partners/pangeo)][PangeoCollaboration]

**For all submissions (brief, 1–2 sentences each):**

* **Target audience & scientific applications:** Researchers, students, librarians, and software authors who need reproducible, standardized references for papers, replication packages, and systematic reviews; OneCite automates multi-source lookup, validation, and formatting.
* **Similar packages & how this differs:** Existing tools are often source-specific or retrieval-only; OneCite provides a general-purpose, multi-database pipeline with metadata completion, de-duplication, and multi-style outputs via CLI and Python API.
* **Pre-submission enquiry:** N/A

* [x] uses an [[OSI approved license](https://opensource.org/licenses)][OsiApprovedLicense].

* [ ] Do you wish to automatically submit to the [[Journal of Open Source Software](http://joss.theoj.org/)][JournalOfOpenSourceSoftware]? **No.**

* [x] I have read the [[author guide](https://www.pyopensci.org/software-peer-review/how-to/author-guide.html)](https://www.pyopensci.org/software-peer-review/how-to/author-guide.html).

---
