### 2025-10-19 885c675
Add OneCite core code
