+++ onecite/pipeline.py
@@ -1,0 +1,3184 @@
+﻿# !/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
+"""OneCite Processing Pipeline Module Implementation of various modules for the 4-stage processing pipeline"""
+
+import re
+import os
+import logging
+import time
+from typing import List, Dict, Optional, Callable, Any
+import requests
+from bs4 import BeautifulSoup
+import bibtexparser
+from thefuzz import fuzz
+from scholarly import scholarly
+from html import unescape
+
+from .core import RawEntry, IdentifiedEntry, CompletedEntry
+from .exceptions import ParseError, ResolverError
+import urllib.parse
+
+
+class ParserModule:
+    """Stage 1: Parse and Extract Module"""
+    
+    def __init__(self):
+        self.logger = logging.getLogger(__name__)
+    
+    def parse(self, input_content: str, input_type: str) -> List[RawEntry]:
+        """Parse input content into a list of raw entries Args: input_content: Input content string input_type: Input type ('txt' or 'bib') Returns: List of raw entries"""
+        self.logger.info(f"Starting to parse {input_type} format input content")
+        
+        if input_type.lower() == 'bib':
+            return self._parse_bibtex(input_content)
+        elif input_type.lower() == 'txt':
+            return self._parse_text(input_content)
+        else:
+            raise ParseError(f"Unsupported input type: {input_type}")
+    
+    def _parse_bibtex(self, bibtex_content: str) -> List[RawEntry]:
+        """Parse BibTeX format content"""
+        entries = []
+        try:
+            bib_database = bibtexparser.loads(bibtex_content)
+            for i, entry in enumerate(bib_database.entries):
+                # Preserve original entry data (Bug #5 fix)
+                # Use dict() to create a proper copy, not just a reference
+                original_entry = dict(entry)
+                
+                # Debug: log what we're preserving
+                self.logger.debug(f"Entry {i}: Preserving original fields: {list(original_entry.keys())}")
+                
+                raw_entry: RawEntry = {
+                    'id': i,
+                    'raw_text': str(entry),
+                    'doi': entry.get('doi'),
+                    'url': entry.get('url'),
+                    'query_string': None,
+                    'original_entry': original_entry  # Store complete original entry
+                }
+                
+                # If no DOI is available, generate query string
+                if not raw_entry['doi']:
+                    query_parts = []
+                    if 'title' in entry:
+                        query_parts.append(entry['title'])
+                    if 'author' in entry:
+                        query_parts.append(entry['author'])
+                    if 'year' in entry:
+                        query_parts.append(entry['year'])
+                    raw_entry['query_string'] = ' '.join(query_parts)
+                
+                entries.append(raw_entry)
+            
+            self.logger.info(f"Successfully parsed {len(entries)} BibTeX entries")
+            return entries
+            
+        except Exception as e:
+            self.logger.error(f"BibTeX parsing failed: {str(e)}")
+            raise ParseError(f"BibTeX parsing failed: {str(e)}")
+    
+    def _parse_text(self, text_content: str) -> List[RawEntry]:
+        """Parse plain text format content"""
+        entries = []
+        
+        # Split text blocks using double newlines
+        text_blocks = text_content.split('\n\n')
+        
+        for i, block in enumerate(text_blocks):
+            block = block.strip()
+            if not block:
+                continue
+            
+            raw_entry: RawEntry = {
+                'id': i,
+                'raw_text': block,
+                'doi': None,
+                'url': None,
+                'query_string': None
+            }
+            
+            # Find DOI (standard format)
+            doi_match = re.search(r'10\.\d{4,}/[^\s,}]+', block)
+            if doi_match:
[…]
