### 2025-10-19 c989dae
Add OneCite core code
