### 2025-10-19 17f382c
Fix test suite: repair conftest.py fixtures and remove Chinese comments
