+++ tests/test_integration.py
@@ -1,0 +1,243 @@
+﻿"""
+Integration tests
+Test complete workflows and examples from README
+"""
+import pytest
+import subprocess
+import os
+import tempfile
+
+class TestIntegration:
+    """Integration tests"""
+
+    def run_onecite_command(self, args, cwd=None):
+        """Helper method to run onecite command"""
+        cmd = ["onecite"] + args
+        try:
+            result = subprocess.run(
+                cmd, 
+                capture_output=True, 
+                text=True, 
+                cwd=cwd,
+                timeout=60  # Increase timeout for integration tests
+            )
+            return result.returncode, result.stdout, result.stderr
+        except subprocess.TimeoutExpired:
+            return -1, "", "Command timed out"
+
+    def test_readme_basic_example(self, create_test_file, temp_dir):
+        """Test basic example from README"""
+        # Example input from README
+        readme_input = """10.1038/nature14539
+
+Attention is all you need
+Vaswani et al.
+NIPS 2017"""
+        
+        test_file = create_test_file(readme_input)
+        output_file = os.path.join(temp_dir, "results.bib")
+        
+        code, stdout, stderr = self.run_onecite_command([
+            "process", test_file, "--output", output_file, "--quiet"
+        ])
+        
+        assert code == 0, f"README example failed: {stderr}"
+        assert os.path.exists(output_file), "Output file should be created"
+        
+        # Verify output content
+        with open(output_file, 'r', encoding='utf-8') as f:
+            content = f.read()
+        
+        # Should contain two entries
+        bib_entries = content.count("@")
+        assert bib_entries >= 1, f"Should contain BibTeX entries, found {bib_entries}"
+        
+        # Check specific content (based on README expectations)
+        content_lower = content.lower()
+        # May contain deep learning related content or attention related content
+        has_relevant_content = any(keyword in content_lower for keyword in [
+            "nature", "deep", "learning", "attention", "vaswani", "nips", "neural"
+        ])
+        assert has_relevant_content, "Output should contain relevant academic content"
+
+    def test_workflow_txt_to_bibtex(self, create_test_file, temp_dir):
+        """Test complete TXT to BibTeX workflow"""
+        # Create input containing different types of references
+        mixed_input = """10.1038/nature14539
+
+1706.03762
+
+Attention is all you need
+Vaswani et al.
+NIPS 2017
+
+https://arxiv.org/abs/1706.03762"""
+        
+        test_file = create_test_file(mixed_input)
+        output_file = os.path.join(temp_dir, "output.bib")
+        
+        code, stdout, stderr = self.run_onecite_command([
+            "process", test_file, 
+            "--input-type", "txt",
+            "--output-format", "bibtex",
+            "--template", "journal_article_full",
+            "--output", output_file,
+            "--quiet"
+        ])
+        
+        assert code == 0, f"TXT to BibTeX workflow failed: {stderr}"
+        assert os.path.exists(output_file), "BibTeX output file should be created"
+        
+        with open(output_file, 'r', encoding='utf-8') as f:
+            content = f.read()
+        
+        assert "@" in content, "Should contain BibTeX entries"
+        assert len(content.strip()) > 100, "Should contain substantial content"
+
+    def test_workflow_bib_to_apa(self, create_test_file, temp_dir):
+        """Test BibTeX to APA workflow"""
+        # Create BibTeX input
+        bib_input = """@article{test2015,
+  title={Deep learning},
+  author={LeCun, Yann and Bengio, Yoshua and Hinton, Geoffrey},
+  journal={Nature},
+  year={2015},
+  volume={521},
+  pages={436--444},
[…]
