### 2025-10-19 198c1e8
Fix test suite: repair conftest.py fixtures and remove Chinese comments
