### 2025-10-19 056a2f8
docs: fix templates.rst to accurately reflect source_priority implementation
