+++ docs/templates.rst
@@ -1,0 +1,373 @@
+Custom Templates
+================
+
+OneCite uses YAML-based templates to define citation metadata field requirements. This guide explains how templates work and how to create custom ones.
+
+Template Basics
+---------------
+
+Templates define the **structure and metadata requirements** for different citation types. OneCite comes with built-in templates for:
+
+- **journal_article_full** - Journal articles with complete metadata
+- **conference_paper** - Conference proceedings papers
+- **book** - Books and monographs
+- **thesis** - Theses and dissertations
+- **software** - Software and code repositories
+- **dataset** - Research datasets
+
+Default Templates Location
+~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+Built-in templates are located in the ``onecite/templates/`` directory:
+
+- ``journal_article_full.yaml``
+- ``conference_paper.yaml``
+- ``book.yaml``
+- ``thesis.yaml``
+- ``software.yaml``
+- ``dataset.yaml``
+
+What Templates Actually Do
+---------------------------
+
+**Important:** OneCite templates define **metadata field requirements**, not output formatting.
+
+Output formats (BibTeX, APA, MLA) are implemented in the Python code, not in the YAML templates.
+
+Templates control:
+
+1. **Which fields** are required or optional for a citation type
+2. **BibTeX entry type** (e.g., @article, @book, @inproceedings)
+3. **Optional field completion** strategy (limited data sources)
+
+Templates DO NOT control:
+
+- Output format style (BibTeX/APA/MLA formatting)
+- Field ordering in the output
+- Punctuation or capitalization rules
+
+Template Structure
+------------------
+
+A template YAML file has three main parts:
+
+1. **name** - Template identifier
+2. **entry_type** - BibTeX entry type (e.g., @article, @book)
+3. **fields** - List of field definitions
+
+Example: Journal Article Template
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+Here's the actual ``journal_article_full.yaml`` template::
+
+    name: journal_article_full
+    entry_type: "@article"
+    fields:
+      - name: author
+        required: true
+      - name: title
+        required: true
+      - name: journal
+        required: true
+      - name: year
+        required: true
+      - name: volume
+        required: false
+        source_priority: 
+          - crossref_api
+          - user_prompt
+      - name: number
+        required: false
+        source_priority:
+          - crossref_api
+          - user_prompt
+      - name: pages
+        required: false
+        source_priority:
+          - crossref_api
+          - google_scholar_scraper
+      - name: publisher
+        required: false
+        source_priority:
+          - crossref_api
+          - user_prompt
+      - name: doi
+        required: false
+        source_priority:
+          - crossref_api
+
+Field Definitions
+~~~~~~~~~~~~~~~~~
+
+Each field has:
+
+- ``name`` - Field name (e.g., author, title, journal)
+- ``required`` - Whether this field is required (true/false)
+- ``source_priority`` - Ordered list of data sources to try (optional fields only)
+
+Available Data Sources for source_priority
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+The ``source_priority`` field controls optional field completion for missing fields:
+
+- ``crossref_api`` - CrossRef API
+- ``google_scholar_scraper`` - Google Scholar scraping (if enabled)
+- ``user_prompt`` - Prompt user to enter manually
+
+Special Citation Types
+~~~~~~~~~~~~~~~~~~~~~~
+
+Some citation types are automatically detected and enriched during identification:
+
+- **Software:** GitHub repositories via GitHub API
+- **Dataset:** Zenodo/Figshare via their APIs
+- **Thesis:** Via external providerRE/BASE APIs
[…]
