### 2025-10-19 0b8ed9e
chore: add pyOpenSci compliance files and configurations
- Add CODE_OF_CONDUCT.md (Contributor Covenant)
- Add CONTRIBUTING.md with development guidelines
- Add CITATION.cff for software citation
- Add .zenodo.json for Zenodo archiving
- Add .readthedocs.yml for documentation hosting
- Add GitHub Actions workflows for CI/CD (tests, docs, publish)
- Add docs/requirements.txt for documentation dependencies
- Update pyproject.toml with additional metadata and URLs
- Add comprehensive Sphinx documentation
