+++ pyproject.toml
@@ -1,0 +1,76 @@
+[build-system]
+requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "onecite"
+version = "0.0.11"
+description = "Universal citation management and academic reference toolkit"
+readme = "README.md"
+requires-python = ">=3.7"
+license = {text = "MIT"}
+authors = [
+    {name = "OneCite Team", email = "team@onecite.org"}
+]
+keywords = ["citation", "bibliography", "bibtex", "academic", "reference", "doi", "arxiv"]
+classifiers = [
+    "Development Status :: 4 - Beta",
+    "Intended Audience :: Science/Research",
+    "Intended Audience :: Developers",
+    "License :: OSI Approved :: MIT License",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.7",
+    "Programming Language :: Python :: 3.8",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Topic :: Software Development :: Libraries :: Python Modules",
+    "Topic :: Text Processing :: Markup",
+]
+
+dependencies = [
+    "requests>=2.25.0",
+    "beautifulsoup4>=4.9.0",
+    "bibtexparser>=1.2.0",
+    "thefuzz>=0.19.0",
+    "scholarly>=1.5.0",
+    "pyyaml>=5.4.0",
+    "mcp>=0.9.0",
+]
+
+[project.optional-dependencies]
+dev = [
+    "pytest>=6.0",
+    "pytest-cov>=2.10",
+    "flake8>=3.8",
+    "black>=20.8b1",
+]
+
+[project.urls]
+Homepage = "https://github.com/HzaCode/OneCite"
+Documentation = "https://onecite.readthedocs.io"
+Repository = "https://github.com/HzaCode/OneCite"
+Issues = "https://github.com/HzaCode/OneCite/issues"
+Changelog = "https://github.com/HzaCode/OneCite/blob/main/docs/changelog.rst"
+"Bug Tracker" = "https://github.com/HzaCode/OneCite/issues"
+
+[project.scripts]
+onecite = "onecite.cli:main"
+onecite-mcp = "onecite_mcp.mcp_server:main"
+
+[tool.setuptools]
+packages = ["onecite", "onecite_mcp"]
+
+[tool.setuptools.package-data]
+onecite = ["templates/*.yaml"]
+
+
+[tool.pytest.ini_options]
+testpaths = ["tests"]
+python_files = ["test_*.py"]
+python_classes = ["Test*"]
+python_functions = ["test_*"]
+
+
+
+
