+++ docs/README.md
@@ -1,0 +1,278 @@
+# OneCite Documentation
+
+Welcome to the OneCite documentation source. This directory contains the reStructuredText (RST) source files for the complete OneCite documentation built with Sphinx.
+
+## Building the Documentation
+
+### Prerequisites
+
+Install the required documentation dependencies:
+
+```bash
+pip install -e ".[dev]"
+# or
+pip install sphinx sphinx-rtd-theme
+```
+
+### Building HTML Documentation
+
+To build the documentation as HTML:
+
+```bash
+# On Linux/macOS
+make html
+
+# On Windows (if make is not available)
+sphinx-build -b html . _build/html
+```
+
+The built documentation will be in `_build/html/`. Open `_build/html/index.html` in your browser to view it.
+
+### Building Other Formats
+
+**PDF:**
+```bash
+make latexpdf
+```
+
+**Man Pages:**
+```bash
+make man
+```
+
+**Texinfo:**
+```bash
+make texinfo
+```
+
+**Epub:**
+```bash
+make epub
+```
+
+### Cleaning Build Files
+
+To remove all built documentation:
+
+```bash
+make clean
+```
+
+## File Structure
+
+```
+docs/
+├── conf.py                      # Sphinx configuration
+├── index.rst                    # Main documentation index
+├── installation.rst             # Installation guide
+├── quick_start.rst              # Quick start guide
+├── basic_usage.rst              # Basic usage guide
+├── advanced_usage.rst           # Advanced usage guide
+├── python_api.rst               # Python API guide
+├── mcp_integration.rst          # MCP integration guide
+├── templates.rst                # Custom templates guide
+├── output_formats.rst           # Output formats guide
+├── faq.rst                      # Frequently asked questions
+├── contributing.rst             # Contributing guidelines
+├── changelog.rst                # Version changelog
+├── api/
+│   ├── core.rst                # Core API reference
+│   ├── exceptions.rst          # Exception reference
+│   └── pipeline.rst            # Pipeline reference
+├── Makefile                     # Build automation
+└── README.md                    # This file
+```
+
+## Documentation Sections
+
+### Getting Started
+- **installation.rst** - How to install OneCite
+- **quick_start.rst** - Get started in 5 minutes
+- **basic_usage.rst** - Basic usage examples
+
+### User Guides
+- **advanced_usage.rst** - Advanced features and techniques
+- **python_api.rst** - Using OneCite as a Python library
+- **mcp_integration.rst** - MCP client integration
+- **templates.rst** - Creating custom citation templates
+- **output_formats.rst** - Understanding output formats
+
+### Reference
+- **api/core.rst** - Core API documentation
+- **api/exceptions.rst** - Exception reference
+- **api/pipeline.rst** - Pipeline processing reference
+
+### Additional
+- **faq.rst** - Frequently asked questions
+- **contributing.rst** - Contribution guidelines
+- **changelog.rst** - Version history and roadmap
+
+## Writing Documentation
+
+### RST Basics
+
+Documentation is written in reStructuredText (RST) format. Here are some basics:
+
+```rst
+Main Section
+============
+
+Subsection
+----------
+
+Subsubsection
+~~~~~~~~~~~~~
+
+**Bold text** and *italic text*
+
+.. code-block:: python
+
+    # Code block
+    import onecite
+    
+    result = onecite.process_references("10.1038/nature14539")
+
+- Bullet list
+- Another item
+  - Nested item
+
+1. Numbered list
+2. Second item
+
+:doc:`Link to file`
+`External link <https://example.com>`_
+```
+
[…]
