+++ .github/workflows/docs.yml
@@ -1,37 +1,39 @@
-name: Documentation
+# Documentation workflow - Currently disabled
+# Uncomment to enable automatic documentation building
 
-on:
-  push:
-    branches: [ main ]
-  pull_request:
-    branches: [ main ]
+# name: Documentation
 
-jobs:
-  build:
-    runs-on: ubuntu-latest
-    
-    steps:
-    - uses: actions/checkout@v3
-    
-    - name: Set up Python
-      uses: actions/setup-python@v4
-      with:
-        python-version: '3.10'
-    
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install sphinx sphinx-rtd-theme
-        pip install -e .
-    
-    - name: Build documentation
-      run: |
-        cd docs
-        sphinx-build -b html . _build/html -W
-    
-    - name: Upload documentation artifacts
-      uses: actions/upload-artifact@v3
-      with:
-        name: documentation
-        path: docs/_build/html
+# on:
+#   push:
+#     branches: [ main ]
+#   pull_request:
+#     branches: [ main ]
 
+# jobs:
+#   build:
+#     runs-on: ubuntu-latest
+#     
+#     steps:
+#     - uses: actions/checkout@v3
+#     
+#     - name: Set up Python
+#       uses: actions/setup-python@v4
+#       with:
+#         python-version: '3.10'
+#     
+#     - name: Install dependencies
+#       run: |
+#         python -m pip install --upgrade pip
+#         pip install sphinx sphinx-rtd-theme
+#         pip install -e .
+#     
+#     - name: Build documentation
+#       run: |
+#         cd docs
+#         sphinx-build -b html . _build/html -W
+#     
+#     - name: Upload documentation artifacts
+#       uses: actions/upload-artifact@v3
+#       with:
+#         name: documentation
+#         path: docs/_build/html
