+++ onecite_mcp/server.py
@@ -1,0 +1,242 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+"""
+OneCite MCP Server
+MCP server for integration with MCP clients like editor client/MCP client
+"""
+
+import json
+import sys
+import asyncio
+from typing import Any, Dict, List, Optional
+import logging
+
+from onecite import process_references, PipelineController
+from onecite.exceptions import OneCiteError
+
+# Configure logging
+logging.basicConfig(level=logging.INFO)
+logger = logging.getLogger(__name__)
+
+
+class OneCiteMCPServer:
+    """OneCite MCP server implementation"""
+    
+    def __init__(self):
+        self.tools = self._register_tools()
+    
+    def _register_tools(self) -> Dict[str, Any]:
+        """Register available tools"""
+        return {
+            "cite": {
+                "description": "Generate academic citations",
+                "parameters": {
+                    "type": "object",
+                    "properties": {
+                        "source": {
+                            "type": "string",
+                            "description": "Source of the literature (DOI, URL, title, etc.)"
+                        },
+                        "style": {
+                            "type": "string",
+                            "description": "Citation format (APA, MLA, Chicago, etc.)",
+                            "default": "APA"
+                        },
+                        "format": {
+                            "type": "string",
+                            "description": "Output format (text, bibtex, json)",
+                            "default": "text"
+                        }
+                    },
+                    "required": ["source"]
+                }
+            },
+            "batch_cite": {
+                "description": "Batch generate citations",
+                "parameters": {
+                    "type": "object",
+                    "properties": {
+                        "sources": {
+                            "type": "array",
+                            "items": {"type": "string"},
+                            "description": "List of literature sources"
+                        },
+                        "style": {
+                            "type": "string",
+                            "description": "Citation format",
+                            "default": "APA"
+                        }
+                    },
+                    "required": ["sources"]
+                }
+            },
+            "search": {
+                "description": "Search academic literature",
+                "parameters": {
+                    "type": "object",
+                    "properties": {
+                        "query": {
+                            "type": "string",
+                            "description": "Search keywords"
+                        },
+                        "limit": {
+                            "type": "integer",
+                            "description": "Result count limit",
+                            "default": 10
+                        }
+                    },
+                    "required": ["query"]
+                }
+            }
+        }
+    
+    async def handle_tool_call(self, tool_name: str, arguments: Dict[str, Any]) -> Dict[str, Any]:
+        """Handle tool calls"""
+        try:
+            if tool_name == "cite":
+                result = await self._cite(
+                    arguments["source"],
+                    arguments.get("style", "APA"),
+                    arguments.get("format", "text")
+                )
+            elif tool_name == "batch_cite":
+                result = await self._batch_cite(
+                    arguments["sources"],
+                    arguments.get("style", "APA")
+                )
+            elif tool_name == "search":
+                result = await self._search(
+                    arguments["query"],
+                    arguments.get("limit", 10)
+                )
+            else:
+                raise ValueError(f"Unknown tool: {tool_name}")
+            
+            return {"success": True, "result": result}
+        except Exception as e:
+            logger.error(f"Tool call failed: {e}")
+            return {"success": False, "error": str(e)}
+    
+    async def _cite(self, source: str, style: str, format: str) -> Any:
+        """Generate a single citation"""
+        try:
+            # Use process_references function to handle citation
+            result = process_references(
+                input_content=source,
+                input_type="txt",
+                output_format=style.lower() if style.lower() in ["apa", "mla", "bibtex"] else "bibtex",
[…]
