+++ requirements.txt
@@ -1,0 +1,11 @@
+requests>=2.25.0
+beautifulsoup4>=4.9.0
+bibtexparser>=1.2.0
+thefuzz>=0.19.0
+scholarly>=1.5.0
+pyyaml>=5.4.0
+mcp>=0.9.0
+
+
+
+

[…]
