### 2025-10-19 1a41f09
docs: add documentation badge and remove coming soon notice
