### 2025-10-19 06c3d15
chore: update contact email to ang@hezhiang.com and add MIT LICENSE
