+++ CITATION.cff
@@ -5,8 +5,7 @@ title: "OneCite: Universal Citation Management and Academic Reference Toolkit"
 authors:
   - family-names: "He"
     given-names: "Zhiang"
-    orcid: "https://orcid.org/0000-0000-0000-0000"
-  - name: "OneCite Team"
+    email: "ang@hezhiang.com"
 repository-code: "https://github.com/HzaCode/OneCite"
 url: "https://github.com/HzaCode/OneCite"
 abstract: "OneCite is a powerful command-line tool and Python library designed to automate the tedious process of citation management. It supports multiple input formats (DOI, arXiv ID, PMID, paper titles, etc.) and generates clean, accurate bibliographic entries in BibTeX, APA, or MLA format."
