### 2025-10-19 bfd5ec5
chore: update contact email to ang@hezhiang.com and add MIT LICENSE
