+++ CONTRIBUTING.md
@@ -15,7 +15,7 @@ Thank you for your interest in contributing to OneCite! This document provides g
 
 ## Code of Conduct
 
-This project adheres to the Contributor Covenant [Code of Conduct](CODE_OF_CONDUCT.md). By participating, you are expected to uphold this code. Please report unacceptable behavior to team@onecite.org.
+This project adheres to the Contributor Covenant [Code of Conduct](CODE_OF_CONDUCT.md). By participating, you are expected to uphold this code. Please report unacceptable behavior to ang@hezhiang.com.
 
 ## Getting Started
 
@@ -264,7 +264,7 @@ If you have questions or need help:
 
 - Open an issue on GitHub
 - Check existing issues and documentation
-- Email team@onecite.org
+- Email ang@hezhiang.com
 
 Thank you for contributing to OneCite! 🎉
 
