### 2025-10-20 d96118f
docs: add CI/Tests badge to README
