+++ README.md
@@ -21,6 +21,7 @@
 
 <div align="center">
 
+[![Tests](https://github.com/HzaCode/OneCite/workflows/Tests/badge.svg?style=flat-square)](https://github.com/HzaCode/OneCite/actions)
 [![Status](https://img.shields.io/badge/Status-Active-green.svg?style=flat-square)](https://github.com/HzaCode/OneCite)
 [![PyPI Version](https://img.shields.io/pypi/v/onecite?color=306998&logo=pypi&style=flat-square)](https://pypi.org/project/onecite/)
 [![Python Version](https://img.shields.io/badge/Python-3.7+-blue?logo=python&style=flat-square)](https://www.python.org)
