### 2025-10-20 91f8fdf
fix: remove non-existent onecite_mcp package from pyproject.toml
