### 2025-10-20 b881e92
Add mock support for CI tests to avoid external API dependencies
