### 2025-10-20 9aca91a
Add mock support for CI tests to avoid external API dependencies
