### 2025-10-20 520a380
Fix mock patches to target correct module paths for CI tests
