### 2025-10-20 cdddbfd
Fix mock patches to target correct module paths for CI tests
