### 2025-10-20 d2db279
Add feedparser dependency to fix arXiv API parsing in CI
