### 2025-10-20 580b13f
update: Fix CI test failures by making assertions more robust
