### 2025-10-20 0b45c28
update: Limit CI tests to Python 3.10+ only
