### 2025-10-20 e7b3053
update: Limit CI tests to Python 3.10+ only
