### 2025-10-20 3f5e7ad
docs: Update Python version requirement to 3.10+ in README
