### 2025-10-28 a561d96
Add MCP badge to README
