### 2025-12-14 118dfc3
pipeline: prioritize DOI in candidate scoring
