### 2025-12-16 4ccc378
chore: pyOpenSci compliance - pre-commit, CHANGELOG, 80% coverage
