+++ tests/test_pipeline_unit.py
@@ -1,0 +1,883 @@
+import builtins
+import json
+import types
+
+import pytest
+import requests
+from unittest.mock import patch
+
+from onecite.pipeline import EnricherModule, FormatterModule, IdentifierModule
+
+
+class DummyResponse:
+    def __init__(self, *, status_code=200, json_data=None, content=b"", headers=None):
+        self.status_code = status_code
+        self._json_data = json_data
+        self.content = content
+        self.headers = headers or {}
+
+    def json(self):
+        return self._json_data
+
+    def raise_for_status(self):
+        if self.status_code >= 400:
+            raise requests.exceptions.HTTPError("HTTP error", response=self)
+
+
+class ImmediateThread:
+    def __init__(self, target=None, args=(), kwargs=None):
+        self._target = target
+        self._args = args
+        self._kwargs = kwargs or {}
+        self.daemon = False
+
+    def start(self):
+        if self._target:
+            self._target(*self._args, **self._kwargs)
+
+
+class NoopThread:
+    def __init__(self, target=None, args=(), kwargs=None):
+        self.daemon = False
+
+    def start(self):
+        return None
+
+
+def test_identifier_extract_github_info():
+    identifier = IdentifierModule()
+
+    def fake_get(url, *args, **kwargs):
+        if url.endswith("/repos/owner/repo"):
+            return DummyResponse(
+                status_code=200,
+                json_data={
+                    "name": "repo",
+                    "description": "desc",
+                    "owner": {"login": "owner"},
+                    "created_at": "2020-01-02T00:00:00Z",
+                    "html_url": "https://github.com/owner/repo",
+                    "language": "Python",
+                    "stargazers_count": 123,
+                },
+            )
+        if url.endswith("/repos/owner/repo/tags"):
+            return DummyResponse(status_code=200, json_data=[{"name": "v1.2.3"}])
+        return DummyResponse(status_code=404, json_data={})
+
+    with patch("onecite.pipeline.requests.get", side_effect=fake_get):
+        info = identifier._extract_github_info("https://github.com/owner/repo")
+
+    assert info is not None
+    assert info["source"] == "github"
+    assert info["repo"] == "owner/repo"
+    assert info["version"] == "1.2.3"
+
+
+def test_identifier_extract_zenodo_info():
+    identifier = IdentifierModule()
+
+    def fake_get(url, *args, **kwargs):
+        if url.endswith("/api/records/12345"):
+            return DummyResponse(
+                status_code=200,
+                json_data={
+                    "metadata": {
+                        "title": "Dataset",
+                        "creators": [{"name": "A"}],
+                        "publication_date": "2021-01-01",
+                        "version": "1.0",
+                        "resource_type": {"type": "dataset"},
+                    }
+                },
+            )
+        return DummyResponse(status_code=404, json_data={})
+
+    with patch("onecite.pipeline.requests.get", side_effect=fake_get):
+        info = identifier._extract_zenodo_info("10.5281/zenodo.12345")
+
+    assert info is not None
+    assert info["source"] == "zenodo"
+    assert info["doi"] == "10.5281/zenodo.12345"
+    assert info["title"] == "Dataset"
+    assert info["authors"] == ["A"]
+
+
+def test_identifier_extract_figshare_info():
+    identifier = IdentifierModule()
+
+    info = identifier._extract_zenodo_info("10.6084/m9.figshare.98765")
+    assert info is not None
+    assert info["source"] == "figshare"
+    assert info["doi"] == "10.6084/m9.figshare.98765"
+
+
+def test_identifier_extract_datacite_info_via_patterns():
+    identifier = IdentifierModule()
+
[…]
