+++ CHANGELOG.md
@@ -1,0 +1,61 @@
+# Changelog
+
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/), and this project adheres to [Semantic Versioning](https://semver.org/).
+
+## [Unreleased]
+
+### Added
+- Comprehensive RST documentation using Sphinx
+- Full API reference documentation
+- FAQ section with common questions
+- Contributing guidelines
+- MCP (Model Context Protocol) integration documentation
+- Pre-commit hooks configuration
+
+### Changed
+- Improved error handling and exception hierarchy
+- Enhanced Python API with better type hints
+- Updated README with more examples
+
+### Fixed
+- Various bug fixes in data source integration
+
+## [0.0.11] - 2024-10-19
+
+### Added
+- MCP server support for MCP client integration
+- Custom YAML-based template system
+- Support for multiple output formats (BibTeX, APA, MLA)
+- Interactive mode for ambiguous reference selection
+- Support for DOI, arXiv, PMID, ISBN, and GitHub identifiers
+- Integration with 9 major academic data sources
+- Comprehensive test suite
+
+### Changed
+- Refactored core processing pipeline
+- Improved data source prioritization
+- Enhanced error messages
+
+### Fixed
+- Fixed encoding issues with special characters
+- Improved robustness of DOI parsing
+- Fixed compatibility with older Python versions
+
+## [0.0.10] - 2024-10-01
+
+### Added
+- Initial Python API
+- Basic citation processing
+- Support for journal articles and conference papers
+
+### Changed
+- Improved search accuracy
+
+### Fixed
+- Fixed several data source API issues
+
+## [0.0.9] and Earlier
+
+See [GitHub Releases](https://github.com/HzaCode/OneCite/releases) for details on older versions.
