+++ CONTRIBUTING.md
@@ -34,7 +34,7 @@ This project adheres to the Contributor Covenant [Code of Conduct](CODE_OF_CONDU
 
 ### Prerequisites
 
-- Python 3.7 or higher
+- Python 3.10 or higher
 - pip
 
 ### Install Development Dependencies
