+++ docs/changelog.rst
@@ -133,9 +133,6 @@ Version History
 
 **Python Support:**
 
-- 3.7+
-- 3.8+
-- 3.9+
 - 3.10+
 - 3.11+
 
