### 2025-12-16 c52db99
chore: pyOpenSci compliance - pre-commit, CHANGELOG, 80% coverage
