+++ tests/test_cli.py
@@ -2,11 +2,16 @@
 Test command line interface functionality
 Test CLI functionality based on README declarations
 """
+import argparse
 import pytest
 import subprocess
 import os
 import tempfile
 from pathlib import Path
+from unittest.mock import Mock, patch
+
+import onecite.cli as cli
+from onecite.exceptions import OneCiteError
 
 class TestCLI:
     """Command line interface tests"""
@@ -76,4 +81,263 @@ class TestCLI:
             "process", test_file, "--output-format", "invalid"
         ])
         assert code != 0, "Should return error for invalid output format"
+
+
+class TestCLIUnit:
+    def test_process_command_missing_input_file(self, capsys):
+        args = argparse.Namespace(
+            command="process",
+            input_file="definitely_not_exists.txt",
+            input_type="txt",
+            template="journal_article_full",
+            output_format="bibtex",
+            output=None,
+            interactive=False,
+            quiet=False,
+        )
+
+        code = cli.process_command(args)
+        captured = capsys.readouterr()
+        assert code == 1
+        assert "Input file not found" in captured.err
+
+    def test_process_command_quiet_and_output_file(self, tmp_path, capsys):
+        input_file = tmp_path / "in.txt"
+        input_file.write_text("10.1038/nature14539", encoding="utf-8")
+        output_file = tmp_path / "out.txt"
+
+        args = argparse.Namespace(
+            command="process",
+            input_file=str(input_file),
+            input_type="txt",
+            template="journal_article_full",
+            output_format="bibtex",
+            output=str(output_file),
+            interactive=False,
+            quiet=True,
+        )
+
+        def fake_process_references(*, input_content, input_type, template_name, output_format, interactive_callback):
+            assert input_content
+            assert input_type == "txt"
+            assert template_name
+            assert output_format == "bibtex"
+            assert interactive_callback([{"title": "T", "authors": [], "journal": "", "year": 2020, "match_score": 75}]) == -1
+            return {"results": ["OK"], "report": {"total": 1, "succeeded": 1, "failed_entries": []}}
+
+        with patch("onecite.cli.process_references", side_effect=fake_process_references):
+            code = cli.process_command(args)
+
+        captured = capsys.readouterr()
+        assert code == 0
+        assert captured.out == ""
+        assert output_file.read_text(encoding="utf-8") == "OK"
+
+    def test_process_command_interactive_selection_prints_report(self, tmp_path, capsys):
+        input_file = tmp_path / "in.txt"
+        input_file.write_text("Some query", encoding="utf-8")
+
+        args = argparse.Namespace(
+            command="process",
+            input_file=str(input_file),
+            input_type="txt",
+            template="journal_article_full",
+            output_format="bibtex",
+            output=None,
+            interactive=True,
+            quiet=False,
+        )
+
+        def fake_process_references(*, input_content, input_type, template_name, output_format, interactive_callback):
+            choice = interactive_callback(
+                [
+                    {"title": "A", "authors": ["X"], "journal": "J", "year": 2020, "match_score": 75},
+                    {"title": "B", "authors": ["Y"], "journal": "J", "year": 2021, "match_score": 74},
+                ]
+            )
+            assert choice == 0
+            return {"results": ["OK"], "report": {"total": 1, "succeeded": 1, "failed_entries": []}}
+
+        with patch("builtins.input", return_value="1"), patch("onecite.cli.process_references", side_effect=fake_process_references):
+            code = cli.process_command(args)
+
+        captured = capsys.readouterr()
+        assert code == 0
+        assert "Found multiple possible matches" in captured.out
+        assert "Processing Report" in captured.out
+        assert "Total entries" in captured.out
+
[…]
