+++ CITATION.cff
@@ -20,5 +20,5 @@ keywords:
   - research
 license: MIT
 version: "0.0.11"
-date-released: "2024-01-01"
+date-released: "2025-08-12"
 

[…]
