+++ .github/workflows/tests.yml
@@ -40,8 +40,10 @@ jobs:
         pytest --cov=onecite --cov-report=xml --cov-report=term
     
     - name: Upload coverage to Codecov
-      uses: codecov/codecov-action@v3
+      uses: codecov/codecov-action@v5
       with:
+        token: ${{ secrets.CODECOV_TOKEN }}
+        slug: HzaCode/OneCite
         file: ./coverage.xml
         flags: unittests
         name: codecov-umbrella
