### 2025-12-16 #267 @HzaCode
Hi @eliotwrobson,

As discussed, this is the submission I would like to move forward with. Please proceed with the pre-review checks for OneCite when you are ready.

Thank you!
