### 2025-12-18 b3399f0
Remove MCP references and fix metadata inconsistencies
