+++ pyproject.toml
@@ -32,7 +32,6 @@ dependencies = [
     "thefuzz>=0.19.0",
     "scholarly>=1.5.0",
     "pyyaml>=5.4.0",
-    "mcp>=0.9.0",
     "feedparser>=6.0.0",
 ]
 
