### 2025-12-18 f983a61
Remove MCP references and fix metadata inconsistencies
