### 2025-12-18 8957fb9
Unify badge styles and reorder
