### 2025-12-19 c8f03c7
Add Awesome CLI Apps badge - Featured in awesome-cli-apps list
