### 2025-12-19 59c2eb9
v0.0.12: Update PyPI badge color and version
