### 2025-12-19 a99cce8
v0.0.12: Update PyPI badge color and version
