+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "onecite"
-version = "0.0.11"
+version = "0.0.12"
 description = "Universal citation management and academic reference toolkit"
 readme = "README.md"
 requires-python = ">=3.10"
