### 2025-12-21 08d480f
Update Awesome CLI Apps badge link to academia section
