### 2025-12-22 eaa1424
fix: synchronize package version with pyproject.toml (0.0.12)
