### 2026-02-01 #267 @HzaCode
Hi @crhea93,

Thank you for the update and for stepping in as editor! I really appreciate your time.

I’m very excited to move forward and see this review as a valuable collaboration to build a truly high-quality tool for the open-source community.

My motivation for OneCite comes from a problem I believe is becoming more critical in the AI era: the need for a deterministic, "ground-truth" toolkit to verify and standardize messy references. As content is generated more automatically, a reliable bridge to turn fragmented metadata from sources like DOIs, GitHub, and arXiv into reproducible citations is more important than ever.

I'm passionate about contributing a robust solution for researchers and developers alike and am looking forward to an interesting and productive review journey. Please let me know if you need anything from my end for the pre-review checks.
