### 2026-02-07 4831cb8
refactor: code review cleanup and test suite improvements
- Restructure test suite with class-based organization and clearer docstrings explaining test strategy and mock design decisions

- Fix DummyResponse mock to properly support streaming responses (raw attribute) matching real requests.Response behavior

- Improve mock_responses router with explicit URL matching and clear fallthrough-to-404 semantics

- Consolidate shared fixtures in conftest.py with descriptive comments on each fixture purpose

- Clean up exception hierarchy and CLI interface

- Update documentation and metadata for cons
[…]
