+++ tests/test_pipeline_unit.py
@@ -1,4 +1,15 @@
+"""
+Unit tests for the three pipeline modules (Identifier → Enricher → Formatter).
+
+Every test here mocks HTTP at the ``requests.get`` level so we never
+touch real APIs.  The helper classes at the top (DummyResponse, ImmediateThread,
+NoopThread) keep the individual test bodies short and readable.
+
+Note: synthetic DOIs like 10.1234/xyz are used intentionally – they don't
+exist on Crossref, which makes it obvious when a mock is missing.
+"""
 import builtins
+import io
 import json
 import types
 
@@ -9,15 +20,22 @@ from unittest.mock import patch
 from onecite.pipeline import EnricherModule, FormatterModule, IdentifierModule
 
 
+# ---------------------------------------------------------------------------
+# Test doubles
+# ---------------------------------------------------------------------------
+
 class DummyResponse:
+    """Minimal ``requests.Response`` stand-in."""
+
     def __init__(self, *, status_code=200, json_data=None, content=b"", headers=None):
         self.status_code = status_code
-        self._json_data = json_data
+        self._json = json_data
         self.content = content
         self.headers = headers or {}
+        self.raw = io.BytesIO(content)
 
     def json(self):
-        return self._json_data
+        return self._json
 
     def raise_for_status(self):
         if self.status_code >= 400:
@@ -25,6 +43,7 @@ class DummyResponse:
 
 
 class ImmediateThread:
+    """Runs the target synchronously in ``start()`` – no real threading."""
     def __init__(self, target=None, args=(), kwargs=None):
         self._target = target
         self._args = args
@@ -37,847 +56,827 @@ class ImmediateThread:
 
 
 class NoopThread:
+    """start() does nothing → simulates a thread that never finishes."""
     def __init__(self, target=None, args=(), kwargs=None):
         self.daemon = False
 
     def start(self):
-        return None
+        pass
 
 
-def test_identifier_extract_github_info():
-    identifier = IdentifierModule()
+# ===================================================================
+# IdentifierModule
+# ===================================================================
 
-    def fake_get(url, *args, **kwargs):
-        if url.endswith("/repos/owner/repo"):
-            return DummyResponse(
-                status_code=200,
-                json_data={
-                    "name": "repo",
-                    "description": "desc",
+class TestIdentifierGitHub:
+
+    def test_extracts_repo_info(self):
+        ident = IdentifierModule()
+
+        def fake_get(url, *a, **kw):
+            if url.endswith("/repos/owner/repo"):
+                return DummyResponse(json_data={
+                    "name": "repo", "description": "desc",
                     "owner": {"login": "owner"},
                     "created_at": "2020-01-02T00:00:00Z",
                     "html_url": "https://github.com/owner/repo",
-                    "language": "Python",
-                    "stargazers_count": 123,
-                },
-            )
-        if url.endswith("/repos/owner/repo/tags"):
-            return DummyResponse(status_code=200, json_data=[{"name": "v1.2.3"}])
-        return DummyResponse(status_code=404, json_data={})
+                    "language": "Python", "stargazers_count": 123,
+                })
+            if url.endswith("/repos/owner/repo/tags"):
+                return DummyResponse(json_data=[{"name": "v1.2.3"}])
+            return DummyResponse(status_code=404, json_data={})
 
-    with patch("onecite.pipeline.requests.get", side_effect=fake_get):
-        info = identifier._extract_github_info("https://github.com/owner/repo")
+        with patch("onecite.pipeline.requests.get", side_effect=fake_get):
[…]
