+++ tests/test_integration.py
@@ -1,166 +1,86 @@
-﻿"""
-Integration tests
-Test complete workflows and examples from README
 """
-import pytest
-import os
-import tempfile
-from unittest.mock import patch
-from .mock_responses import mock_requests_get
+End-to-end workflow tests – the kind of thing you'd manually try in a
+terminal to convince yourself the tool actually works.
+
+All API calls are mocked via the ``run_onecite_process`` fixture.
+"""
+
 
 class TestIntegration:
-    """Integration tests"""
 
-    def run_onecite_process(self, input_content, input_type="txt", output_format="bibtex", template="journal_article_full"):
-        """Helper method to run onecite processing with mocked API calls"""
-        try:
-            # Mock requests.get in all modules to avoid real API calls
-            with patch('onecite.pipeline.requests.get', side_effect=mock_requests_get), \
-                 patch('onecite.core.requests.get', side_effect=mock_requests_get), \
-                 patch('requests.get', side_effect=mock_requests_get):
-                from onecite import process_references
-                
-                def mock_callback(candidates):
-                    return 0  # Always choose the first candidate
-                
-                result = process_references(
-                    input_content=input_content,
-                    input_type=input_type,
-                    template_name=template,
-                    output_format=output_format,
-                    interactive_callback=mock_callback
-                )
-                
-                # Simulate CLI output
-                output_content = '\n\n'.join(result['results'])
-                return 0, output_content, "", result
-        except Exception as e:
-            return 1, "", str(e), None
+    @staticmethod
+    def _check(result):
+        """Common assertions on the result dict."""
+        assert isinstance(result, dict)
+        assert "report" in result and "results" in result
 
-    def test_readme_basic_example(self):
-        """Test basic example from README"""
-        # Example input from README
-        readme_input = """10.1038/nature14539
-
-Attention is all you need
-Vaswani et al.
-NIPS 2017"""
-        
-        code, stdout, stderr, result = self.run_onecite_process(readme_input)
-        
-        assert code == 0, f"README example failed: {stderr}"
-        
-        # Should contain entries or at least attempt processing
-        assert result is not None, "Should return result"
-        assert result['report']['total'] >= 1, "Should process at least one entry"
-
-    def test_workflow_txt_to_bibtex(self):
-        """Test complete TXT to BibTeX workflow"""
-        # Create input containing different types of references
-        mixed_input = """10.1038/nature14539
-
-1706.03762"""
-        
-        code, stdout, stderr, result = self.run_onecite_process(
-            mixed_input, input_type="txt", output_format="bibtex"
+    def test_readme_example(self, run_onecite_process):
+        """The exact snippet shown in the README."""
+        code, _, err, result = run_onecite_process(
+            "10.1038/nature14539\n\nAttention is all you need\nVaswani et al.\nNIPS 2017"
         )
-        
-        assert code == 0, f"TXT to BibTeX workflow failed: {stderr}"
-        assert result is not None, "Should return result"
+        assert code == 0, err
+        self._check(result)
+        assert result["report"]["total"] >= 1
 
-    def test_workflow_bib_to_apa(self):
-        """Test BibTeX to APA workflow"""
-        # Create BibTeX input
-        bib_input = """@article{test2015,
-  title={Deep learning},
-  author={LeCun, Yann and Bengio, Yoshua and Hinton, Geoffrey},
-  journal={Nature},
-  year={2015},
-  volume={521},
-  pages={436--444},
-  doi={10.1038/nature14539}
-}"""
-        
[…]
