+++ tests/conftest.py
@@ -1,37 +1,47 @@
-﻿"""
-Pytest configuration and fixtures for OneCite tests.
 """
-import pytest
-import sys
+Shared fixtures for the OneCite test suite.
+
+Most tests that hit external APIs (Crossref, arXiv, Semantic Scholar …)
+go through the ``run_onecite_process`` fixture which patches
+``requests.get`` with our hand-crafted mock responses so the CI stays
+deterministic and fast.
+"""
 import os
-import tempfile
+import sys
 import shutil
+import tempfile
 
-# Add parent directory to path for imports
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+import pytest
+from unittest.mock import patch
 
+from .mock_responses import mock_requests_get
+
+# Make sure the repo root is importable even when pytest is invoked
+# from an unexpected working directory.
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
+
+
+# ---------------------------------------------------------------------------
+# Simple scalar fixtures
+# ---------------------------------------------------------------------------
 
 @pytest.fixture
 def sample_doi():
-    """Sample DOI for testing."""
     return "10.1038/nature14539"
 
 
 @pytest.fixture
 def sample_arxiv():
-    """Sample arXiv ID for testing."""
     return "1706.03762"
 
 
 @pytest.fixture
 def sample_text_query():
-    """Sample text query for testing."""
     return "Attention is all you need, Vaswani et al., NIPS 2017"
 
 
 @pytest.fixture
 def sample_bibtex():
-    """Sample BibTeX entry for testing."""
     return """@article{sample2020,
   title={Sample Article},
   author={Smith, John and Doe, Jane},
@@ -44,39 +54,85 @@ def sample_bibtex():
 
 @pytest.fixture
 def sample_references():
-    """Sample references for testing - returns a dictionary."""
+    """Dict of canned inputs keyed by type – keeps individual tests short."""
     return {
         "doi_only": "10.1038/nature14539",
         "arxiv_id": "1706.03762",
         "conference_paper": "Attention is all you need\nVaswani et al.\nNIPS 2017",
-        "bibtex_entry": """@article{sample2020,
-  title={Sample Article},
-  author={Smith, John and Doe, Jane},
-  journal={Nature},
-  year={2020},
-  volume={580},
-  pages={1-10}
-}"""
+        "bibtex_entry": (
+            "@article{sample2020,\n"
+            "  title={Sample Article},\n"
+            "  author={Smith, John and Doe, Jane},\n"
+            "  journal={Nature},\n"
+            "  year={2020},\n"
+            "  volume={580},\n"
+            "  pages={1-10}\n"
+            "}"
+        ),
     }
 
 
+# ---------------------------------------------------------------------------
+# File / directory helpers
+# ---------------------------------------------------------------------------
+
 @pytest.fixture
 def create_test_file(tmp_path):
-    """Factory fixture to create test files."""
-    def _create_file(content, filename="test_input.txt"):
-        """Create a test file with given content."""
-        file_path = tmp_path / filename
-        file_path.write_text(content, encoding='utf-8')
-        return str(file_path)
-    
-    return _create_file
+    """Return a factory that writes *content* to a temp file and gives back
+    its path as a string (most CLI helpers expect ``str``, not ``Path``)."""
+
+    def _create(content, filename="test_input.txt"):
+        p = tmp_path / filename
+        p.write_text(content, encoding="utf-8")
+        return str(p)
+
+    return _create
 
 
 @pytest.fixture
 def temp_dir():
-    """Create a temporary directory for testing."""
[…]
