+++ onecite/pipeline.py
@@ -1,7 +1,7 @@
-﻿# !/usr/bin/env python3
+# !/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
-"""OneCite Processing Pipeline Module Implementation of various modules for the 4-stage processing pipeline"""
+"""OneCite 4-stage processing pipeline."""
 
 import re
 import os
@@ -12,12 +12,27 @@ import requests
 from bs4 import BeautifulSoup
 import bibtexparser
 from thefuzz import fuzz
-from scholarly import scholarly
+try:
+    from scholarly import scholarly
+except ImportError:
+    scholarly = None
 from html import unescape
 
 from .core import RawEntry, IdentifiedEntry, CompletedEntry
 from .exceptions import ParseError, ResolverError
 import urllib.parse
+
+
+def _safe_year(date_obj):
+    """Safely extract year from a CrossRef date object like {'date-parts': [[2015, 3, 1]]}."""
+    if not date_obj:
+        return None
+    parts = date_obj.get('date-parts', [])
+    if parts and isinstance(parts, list) and len(parts) > 0:
+        inner = parts[0]
+        if isinstance(inner, list) and len(inner) > 0:
+            return inner[0]
+    return None
 
 
 class ParserModule:
@@ -27,7 +42,7 @@ class ParserModule:
         self.logger = logging.getLogger(__name__)
     
     def parse(self, input_content: str, input_type: str) -> List[RawEntry]:
-        """Parse input content into a list of raw entries Args: input_content: Input content string input_type: Input type ('txt' or 'bib') Returns: List of raw entries"""
+        """Parse input content into raw entries."""
         self.logger.info(f"Starting to parse {input_type} format input content")
         
         if input_type.lower() == 'bib':
@@ -43,12 +58,9 @@ class ParserModule:
         try:
             bib_database = bibtexparser.loads(bibtex_content)
             for i, entry in enumerate(bib_database.entries):
-                # Preserve original entry data (Bug #5 fix)
-                # Use dict() to create a proper copy, not just a reference
                 original_entry = dict(entry)
                 
-                # Debug: log what we're preserving
-                self.logger.debug(f"Entry {i}: Preserving original fields: {list(original_entry.keys())}")
+                self.logger.debug(f"Entry {i}: original fields: {list(original_entry.keys())}")
                 
                 raw_entry: RawEntry = {
                     'id': i,
@@ -56,7 +68,7 @@ class ParserModule:
                     'doi': entry.get('doi'),
                     'url': entry.get('url'),
                     'query_string': None,
-                    'original_entry': original_entry  # Store complete original entry
+                    'original_entry': original_entry
                 }
                 
                 # If no DOI is available, generate query string
@@ -99,10 +111,9 @@ class ParserModule:
                 'query_string': None
             }
             
-            # Find DOI (standard format)
             doi_match = re.search(r'10\.\d{4,}/[^\s,}]+', block)
             if doi_match:
-                raw_entry['doi'] = doi_match.group()
+                raw_entry['doi'] = doi_match.group().rstrip('.,;:)]')
             else:
                 # Try to find article ID patterns that might be convertible to DOI
                 # Common patterns: e0000429, PMC123456, etc.
@@ -112,11 +123,9 @@ class ParserModule:
                     # Note potential PLOS article ID but don't assume specific journal
                     # Let Cross resolve the actual DOI during identification
                     self.logger.info(f"Entry {i} found potential PLOS article ID {article_id}, will attempt resolution via CrossRef")
-                    # Add to query string instead of assuming DOI
                     if not raw_entry['query_string']:
                         raw_entry['query_string'] = block
             
-            # Find URL
             url_match = re.search(r'https?://[^\s]+', block)
             if url_match:
                 raw_entry['url'] = url_match.group()
[…]
