+++ tests/__init__.py
@@ -1,7 +1,5 @@
 """
 OneCite Test Suite
-
-This package contains comprehensive tests for the OneCite library.
 """
 
 
