### 2026-02-07 d1061f8
fix: mock scholarly module object in CI where scholarly is not installed
The scholarly package is an optional dependency. When not installed, the module-level variable is None, so patching scholarly.search_pubs directly raises AttributeError. Use patch.object on the pipeline module instead.
