### 2026-02-07 131991a
Add disclosure section to README
Added disclosure about AI assistance in development.
