### 2026-02-07 #267 @HzaCode
Hi @crhea93,

Thank you. I have completed the onboarding survey.

Regarding the Generative AI disclosure, I used Generative AI tools to **assist** in the development process. Since I utilize them iteratively to streamline implementation details and refine documentation, it is difficult to isolate specific locations as their usage is blended into the general workflow.

However, all AI-assisted content was manually reviewed, tested, and integrated by the maintainer.Crucially, I confirm that no LLMs are used by the package at runtime.

I have added a corresponding note to the `README.md`.
