### 2026-02-08 #267 @crhea93
Alright thank you for putting that together! I've already found you an editor, so I will link them presently.
