### 2026-02-09 82c9964
ci: auto-deploy docs to gh-pages on push to main
