+++ onecite_mcp/__init__.py
@@ -1,10 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
-"""
-OneCite MCP (Model Context Protocol) Server
-MCP server implementation for MCP client integration
-"""
-
-__version__ = "0.0.4"
-__author__ = "OneCite Team"
