### 2026-02-09 12b1dea
release: v0.1.0 - documentation cleanup and PyOpenSci review preparation
