+++ docs/installation.rst
@@ -4,7 +4,7 @@ Installation Guide
 System Requirements
 -------------------
 
-- Python 3.7 or higher
+- Python 3.10 or higher
 - pip package manager
 - Internet connection (for accessing data sources)
 
