+++ .github/workflows/publish.yml
@@ -12,10 +12,10 @@
 #     runs-on: ubuntu-latest
 #     
 #     steps:
-#     - uses: actions/checkout@v3
+#     - uses: actions/checkout@v4
 #     
 #     - name: Set up Python
-#       uses: actions/setup-python@v4
+#       uses: actions/setup-python@v5
 #       with:
 #         python-version: '3.10'
 #     

[…]
