### 2026-02-09 3b53575
release: v0.1.0 - documentation cleanup and PyOpenSci review preparation
