+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "onecite"
-version = "0.0.12"
+version = "0.1.0"
 description = "Citation management and academic reference toolkit"
 readme = "README.md"
 requires-python = ">=3.10"
@@ -47,7 +47,7 @@ dev = [
 
 [project.urls]
 Homepage = "https://github.com/HzaCode/OneCite"
-Documentation = "https://onecite.readthedocs.io"
+Documentation = "https://hezhiang.com/OneCite/"
 Repository = "https://github.com/HzaCode/OneCite"
 Issues = "https://github.com/HzaCode/OneCite/issues"
 Changelog = "https://github.com/HzaCode/OneCite/blob/main/docs/changelog.rst"
