+++ .readthedocs.yml
@@ -1,38 +1,0 @@
-# Read the Docs configuration file
-# See https://docs.readthedocs.io/en/stable/config-file/v2.html for details
-
-version: 2
-
-# Build documentation in the docs/ directory with Sphinx
-sphinx:
-  configuration: docs/conf.py
-  fail_on_warning: false
-
-# Build documentation with MkDocs
-#mkdocs:
-#  configuration: mkdocs.yml
-
-# Optionally build your docs in additional formats such as PDF and ePub
-formats:
-  - pdf
-  - epub
-
-# Set the version of Python and requirements required to build your docs
-build:
-  os: ubuntu-22.04
-  tools:
-    python: "3.10"
-  jobs:
-    post_create_environment:
-      - pip install --upgrade pip setuptools wheel
-    post_install:
-      - pip install -e .
-
-# Install Python dependencies
-python:
-  install:
-    - requirements: requirements.txt
-    - requirements: docs/requirements.txt
-    - method: pip
-      path: .
-
