+++ CITATION.cff
@@ -19,6 +19,6 @@ keywords:
   - arxiv
   - research
 license: MIT
-version: "0.0.11"
-date-released: "2025-08-12"
+version: "0.1.0"
+date-released: "2025-02-09"
 
