### 2026-02-11 #267 @slobentanzer
## Editor response to review:

---

## Editor comments

Hi @HzaCode, many thanks for the submission, looks like a very cool package / purpose! I have started to look for reviewers, already got one response 🤘

- [ ] reviewer 3 (if applicable)

Please get in touch with any questions or concerns!
---

Reviewers:
- @giovana-morais
- @sneakers-the-rat

Due date: 2026-02-25
