### 2026-02-11 #267 @sneakers-the-rat
Howdy there, your friendly neighborhood reviewer 2 checking in. I am going to follow the usual JOSS pattern for software peer review, where because github doesn't have threading in issues, i'm going to raise issues in the target repository and link them back here so that review observations can be addressed as they're raised without muddling this issue up too badly :)

- [x] As the reviewer, I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).

- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience in the README file.
- [x] **Short quickstart tutorials** demonstrating significant functionality that successfully runs locally.
- [x] **Metadata** including author(s), author e-mail(s), a URL, and any other relevant metadata, for example, in a `pyproject.toml` file or elsewhere.

*NOTE: If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be wider than high. A badge for pyOpenSci peer review will be provided when the package is accepted.*

The package structure should follow the general community best practices. In general, please consider whether:

- 2026-02-11: 5h

---

### Review Issues

- https://github.com/HzaCode/OneCite/issues/5
- https://github.com/HzaCode/OneCite/issues/6
-  https://github.com/HzaCode/OneCite/issues/7
-  https://github.com/HzaCode/OneCite/issues/8
-  https://github.com/HzaCode/OneCite/issues/9
-  https://github.com/HzaCode/OneCite/issues/10
-  https://github.com/HzaCode/OneCite/issues/11
-  https://github.com/HzaCode/OneCite/issues/12
-  https://github.com/HzaCode/OneCite/issues/13
-  https://github.com/HzaCode/OneCite/issues/14
-  https://github.com/HzaCode/OneCite/issues/15
-  https://github.com/HzaCode/OneCite/issues/16
-  https://github.com/HzaCode/OneCite/issues/17
-  https://github.com/HzaCode/OneCite/issues/18
-  https://github.com/HzaCode/OneCite/issues/19
-  https://github.com/HzaCode/OneCite/issues/20
-  https://github.com/HzaCode/OneCite/issues/21
-  https://github.com/HzaCode/OneCite/issues/22
-  https://github.com/HzaCode/OneCite/issues/23
-  https://github.com/HzaCode/OneCite/issues/24
-  https://github.com/HzaCode/OneCite/issues/25
-  https://github.com/HzaCode/OneCite/issues/26
-  https://github.com/HzaCode/OneCite/issues/27
-  https://github.com/HzaCode/OneCite/issues/29
-  https://github.com/HzaCode/OneCite/issues/30
-  https://github.com/HzaCode/OneCite/issues/31
-  https://github.com/HzaCode/OneCite/issues/32
-  https://github.com/HzaCode/OneCite/issues/34

This package is a fuzzy matcher that takes freeform strings as inputs and outputs bibtex citations by querying a set of metadata sources. It runs and indeed produces outputs, but the outputs are usually wrong unless the input contains a DOI or the title is completely unique. That the package works is pretty much a testament to the fact that crossref works.

It's not altogether clear to me why this tool is needed - I can get mostly the same results by just querying crossref or openalex directly, and I would likely just use zotero to resolve a DOI or title into metadata. That's fine to me, I don't have to understand the motivation of everything, but I do understand the structure of the problem, and it should be pretty simple: there's some input query string, we do some initial preprocessing to try and recognize common patterns, we either find an exact match (e.g. from a DOI or ISBN) or else we query available sources until we find something that looks like one, spit that out to bibtex, and done. This package has managed to make that relatively straightforward and linear process into something that's very challenging to follow and understand, and every part of that pipeline seems to happens everywhere. The core part of that operation - the fuzzy matching part - is the weakest part of the process, and in many cases i was unable to get the correct result no matter how many exactly matching metadata fields i added.

I have done a decent bit of software peer review in my time, and I love the process of trying to collaboratively improve a package - no package is ever completed, the whole purpose is to get external eyes on it so that things that might be invisible to the author can be improved, knowledge can be shared among peers, and so on. This package is a good example of how LLMs break the underlying social covenant of software development that makes that assumption of goodwill difficult. I have also taught many new programmers and am no stranger to patterns of errors, but even in newbie code, the errors usually have some plausible explanation. Much of the structure of this package does not make sense to me. The code is extraordinarily verbose, with thousands of fallback conditions, duplicated logic, and I'm not sure what it would take to get this package to a state that it could be a reliable, trusted piece of research software short of a full rewrite. While ordinarily I would be fine working with the authors, teaching techniques, and usually even try and contribute a few pull requests myself, since the package seems to have been largely LLM generated, I have no reason to believe my time would be well spent - in some sections, the only explanation to me for how the code could be the way that it is was if I were the first person to have read it. Since I can't know whether the authors have spent their share of time writing the package, I am hesitant to spend more of my time trying to improve it, particularly since it is very likely that rather than having a discussion as peers, all the time I have already spent writing issues will just be dumped into the LLM, some trivial code jittering will happen, but the package won't be improved at a design and architecture level.

That being said, however, I'm happy to engage in continued dialogue with the authors (please write in your natural voice, translation is completely fine if english is not your first language, I would much rather read your voice than have a conversation with an LLM) about what can be done, but I think this package would take substantial work to reach a standard where I think it adequately delivers what it promises. I don't like being nitpicky, my goal is just to make sure that if my name is on something confirming i've reviewed it that it is at least plausible. Who knows, maybe we can have a productive human-to-human mutually beneficial review and learning process in the era of LLMs, i'm willing to try and find out.
