### 2026-02-11 HzaCode/OneCite#6 @sneakers-the-rat
[pyos][docs] Data sources inaccurate
Part of: https://github.com/pyOpenSci/software-submission/issues/267#issuecomment-3887402196

The [README](https://github.com/HzaCode/OneCite/blob/12b1dea45a2b7ddcc60b1abc6cc29984b1aefbc8/README.md)

lists these data sources:

- CrossRef
- Semantic Scholar
- OpenAlex
- PubMed
- dblp
- arXiv
- DataCite
- Zenodo
- Google Books

However I find no mechanism in the code for these:

- OpenAlex
- dblp
