### 2026-02-11 HzaCode/OneCite#7 @sneakers-the-rat
[pyos][meta] Missing statement of need
Part of: https://github.com/pyOpenSci/software-submission/issues/267#issuecomment-3887402196

Submissions require a statement of need ( https://www.pyopensci.org/software-peer-review/how-to/reviewer-guide.html )

However I do not find one in the README.md or the docs

Please add a statement of need that describes the unmet need that this package intends to address. this is important to understand the intended scope and application of the software - in this case it seems like it shoudl be straightforward: unstructured text to citations, but the statement of need anchors expectations and gives reviewers a concrete target to assess the package against.
