### 2026-02-11 HzaCode/OneCite#8 @sneakers-the-rat
[pyos][docs] "content type" selection appears inaccurate
Part of: https://github.com/pyOpenSci/software-submission/issues/267#issuecomment-3887402196

The documentation (https://github.com/HzaCode/OneCite/blob/12b1dea45a2b7ddcc60b1abc6cc29984b1aefbc8/docs/advanced_usage.rst?plain=1#L51-L76 ) reads:

> OneCite routes queries to data sources based on content type:
>
> For Biomedical Literature
>
> Add search terms related to medicine, biology, or health:
>
> onecite process medical_refs.txt
>
> This will prioritize PubMed when available.
>
> For Computer Science
>
>Add search terms related to CS topics:
>
> onecite process cs_refs.txt
>
> This will prioritize DBLP and arXiv.
>
> For General Academic Work
>
> Mixed references will use CrossRef and Semantic Scholar:
>
> onecite process general_refs.txt

This appears to suggest that the filename used when defining input references affects the program operation somehow. However i find no mechanism for that.

The file is read here: https://github.com/HzaCode/OneCite/blob/12b1dea45a2b7ddcc60b1abc6cc29984b1aefbc8/onecite/cli.py#L125

and the filename is never referenced again
