### 2026-02-11 HzaCode/OneCite#9 @sneakers-the-rat
[pyos][docs] `--input-type` not mentioned in "converting between formats"
Part of: https://github.com/pyOpenSci/software-submission/issues/267#issuecomment-3887402196

User is shown an example of giving input in `.bib` format: https://github.com/HzaCode/OneCite/blob/12b1dea45a2b7ddcc60b1abc6cc29984b1aefbc8/docs/advanced_usage.rst?plain=1#L130

however without explicitly passing [`--input-type bib`](https://github.com/HzaCode/OneCite/blob/12b1dea45a2b7ddcc60b1abc6cc29984b1aefbc8/onecite/cli.py#L73), the input is parsed as text: https://github.com/HzaCode/OneCite/blob/12b1dea45a2b7ddcc60b1abc6cc29984b1aefbc8/onecite/pipeline.py#L61
