### 2026-02-11 HzaCode/OneCite#15 @sneakers-the-rat
[pyos][style] Package claims to use `black` formatting but it doesn't
Part of: https://github.com/pyOpenSci/software-submission/issues/267#issuecomment-3887402196

Package claims to use black for formatting: https://github.com/HzaCode/OneCite/blob/12b1dea45a2b7ddcc60b1abc6cc29984b1aefbc8/CONTRIBUTING.md?plain=1#L159

however if i run `black .` the entire everything is reformatted.
