### 2026-02-11 HzaCode/OneCite#16 @sneakers-the-rat
[pyos][design] Templating system conflicts with package design
Part of: https://github.com/pyOpenSci/software-submission/issues/267#issuecomment-3887402196

there is a templating system (that isn't *really* a templating system, it's more of a config system?) that describe which fields are required, and which sources should be checked for them. Templates are applied per pipeline run

The package is designed such that someone can put a collection of reference query strings in whatever format in a text document and get a `.bib` file on the other side.

It thus doesn't make sense to have a single template per pipeline run, the only way that would work is if all items in the input are of the same type, however this isn't discussed in the docs, and the examples are mostly mixed type.

The output i'm getting seems to ignore the `entry_type` anyway, as there is a *ton* of logic in the code to detect and change this.

It doesn't make sense to have templates like this *anyway,* since once a work is identified we should just take the entirety of the metadata returned by the canonical metadata source (or have some structured means of completing the metadata).

I'm not really sure what is to be done here, it's not entirely clear what purpose the templates serve aside from being a list of fields to be queried, but the field querying logic doesn't make much sense either (will discuss in followup issue). It seems like either the docs should clarify that all the items in an input list should be of the same type, or the matching logic in the code should be used to select a template.
