### 2026-02-12 HzaCode/OneCite#19 @sneakers-the-rat
[pyos][bug] One hardcoded "well known paper" masks failure of example input used in docs
Part of: https://github.com/pyOpenSci/software-submission/issues/267#issuecomment-3887402196

There's one hardcoded paper for "attention is all you need" - https://github.com/HzaCode/OneCite/blob/12b1dea45a2b7ddcc60b1abc6cc29984b1aefbc8/onecite/pipeline.py#L189-L200

The comment says "that might not have DOIs" but actually that paper absolutely does have a DOI by virtue of being on arXiv, the arxiv link is right there!

If that is removed, then the result that's returned for the input "attention is all you need" (which is used in multiple places in the docs as an example) is this

```bibtex
@inproceedings{Choi2020Channel,
  doi = "10.1609/AAAI.V34I07.6693",
  title = "Channel Attention Is All You Need for Video Frame Interpolation",
  author = "Choi, Myungsub and Kim, Heewon and Han, Bohyung and Xu, Ning and Lee, Kyoung Mu",
  journal = "Proceedings of the AAAI Conference on Artificial Intelligence",
  year = 2020,
  volume = 34,
  number = 07,
  pages = "10663-10671",
  publisher = "Association for the Advancement of Artificial Intelligence (AAAI)",
  url = "https://doi.org/10.1609/aaai.v34i07.6693",
  type = "journal-article",
}
```

needless to say there should be *no* hardcoded metadata, especially for inputs used as examples in the documentation.
