### 2026-02-12 HzaCode/OneCite#21 @sneakers-the-rat
[pyos][bug] Crossref API requests shoudl have `mailto` and a User Agent that identifies the package
Part of: https://github.com/pyOpenSci/software-submission/issues/267#issuecomment-3887402196

See the crossref api docs: https://www.crossref.org/documentation/retrieve-metadata/rest-api/access-and-authentication/

Users of the crossref API should identify themselves. this also increases the rate and concurrency limits.
