### 2026-02-12 HzaCode/OneCite#25 @sneakers-the-rat
[pyos][bug] Semantic Scholar returns 429 most of the time
Part of: https://github.com/pyOpenSci/software-submission/issues/267#issuecomment-3886731422

When processing a reference list with more than ~3 items, semantic scholar just always returns a 429 from ratelimiting.

```
$ onecite process references.txt -o results.bib
Semantic Scholar returned status code 429
Semantic Scholar returned status code 429
Semantic Scholar returned status code 429
Semantic Scholar returned status code 429
```
