### 2026-02-12 #267 @giovana-morais
hello! i'm reviewer 1 and i'll follow the same procedure as 2 @sneakers-the-rat and link any issues of the repo directly here.

---

- [x] As the reviewer, I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).

- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience in the README file.
- [x] **Short quickstart tutorials** demonstrating significant functionality that successfully runs locally.
- [x] **Metadata** including author(s), author e-mail(s), a URL, and any other relevant metadata, for example, in a `pyproject.toml` file or elsewhere.

*NOTE: If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be wider than high. A badge for pyOpenSci peer review will be provided when the package is accepted.*

The package structure should follow the general community best practices. In general, please consider whether:

Estimated hours spent reviewing: 3h

---

So this package is a fuzzy finder for references. In theory, it accepts different formats for different references and then, based on that disorganized list of references, it outputs a single .bib file with everything.
From my personal perspective and experience, I do not see how this package would be different from searching the internet or using a reference manager like Zotero, Mendeley or even exporting from Google Scholar lists. The motivation for using this package is unclear. Even if we changed it to accept strings as input instead of a file, as I [suggested here](https://github.com/HzaCode/OneCite/issues/36), it would still be equivalent of doing an internet search. All the tests I ran that were not in the examples returned wrong results, even with the interactive mode, that is supposed do disambiguate between matching results.

@sneakers-the-rat covered pretty much all of the code issues, so I won't be discussing them. I do want to mention that the codebase seems unnecessarily complicated and core things do not make sense, such as the ranking when you have ambiguous matches. I do not think that this code is ready to publish given the amount of bugs and unreliable behavior.
