### 2026-02-12 HzaCode/OneCite#29 @sneakers-the-rat
[pyos][bug] `_complete_fields` is a no-op by default and defeats the purpose of templates
Part of: https://github.com/pyOpenSci/software-submission/issues/267#issuecomment-3886731422

During enrichment, if we are missing fields that the template suggests we should have, we attempt to fill them in by querying one of several sources, eventually ending up here: ﻿﻿https://github.com/HzaCode/OneCite/blob/12b1dea45a2b7ddcc60b1abc6cc29984b1aefbc8/onecite/pipeline.py#L2865-L2883

Two of those just `continue`, and the other is google scholar, which is disabled by default with no way to enable it from the CLI. So that method is a constitutive No-op from the CLI. Since the entire purpose of the templates are to specify the fields we should have, and since the method for filling in missing fields is a no-op, the templates are pointless.

The structure of that is also bizarre - it doesn't make sense to query google scholar multiple times, once for each field, and the threading is implemented incorrectly without `join`ing the thread.
