### 2026-02-12 HzaCode/OneCite#31 @sneakers-the-rat
[pyos][bug] APA citation generator is incorrect
Part of: https://github.com/pyOpenSci/software-submission/issues/267#issuecomment-3886731422

See: https://owl.purdue.edu/owl/research_and_citation/apa_style/apa_formatting_and_style_guide/reference_list_basic_rules.html

Essentially none of the rules of APA are followed.

This is what it actually takes to format something as an APA citation - https://github.com/citation-style-language/styles/blob/master/apa.csl

this feature should likely be removed. just export bibtex and import that into zotero and let it do the formatting.
