### 2026-02-12 HzaCode/OneCite#32 @sneakers-the-rat
[pyos][bug] MLA citation format is incorrect
Part of: https://github.com/pyOpenSci/software-submission/issues/267#issuecomment-3886731422

Again, this is what it takes to format something as MLA: https://github.com/citation-style-language/styles/blob/master/modern-language-association.csl

The simplest examples sort of work, but there's lots of bugs here like double periods, capitalization, number formatting, etc.

As with #31 it's not really clear to me why this package would attempt to output formatted references, bibtex makes sense though.
