### 2026-02-12 HzaCode/OneCite#33 @giovana-morais
[pyos][docs] practical examples should be self-contained
part of https://github.com/pyOpenSci/software-submission/issues/267#issuecomment-3888141858

in the [practical examples](https://hezhiang.com/OneCite/basic_usage.html#practical-examples) section of the documentation, it would be very useful to have simple files (1 or 2 entries each) to showcase what the library is actually doing.

for example, it is unclear what it means to "enhance" a bib entry. i'm assuming that it is completing the missing fields, but i cannot ensure that this is the case because it is the first time "enhancement" appears.
