### 2026-02-12 HzaCode/OneCite#34 @sneakers-the-rat
[pyos][tests] Most tests don't test anything
Part of: https://github.com/pyOpenSci/software-submission/issues/267#issuecomment-3886731422

E.g. here: https://github.com/HzaCode/OneCite/blob/12b1dea45a2b7ddcc60b1abc6cc29984b1aefbc8/tests/test_input_formats.py#L26-L36

allegedly we are testing whether these are correctly parsed as DOIs, but it tests nothing of the kind. It just asserts that the processing didn't fail - which it arguably *should* since the network requests that would get the metadata are mocked. This whole test module is like this, as are many of the other test modules.

e.g. here we *don't* mock network, but we still don't test anything aside from "some bibtex was generated" - this does not test whether the input DOI was correctly parsed: https://github.com/HzaCode/OneCite/blob/12b1dea45a2b7ddcc60b1abc6cc29984b1aefbc8/tests/test_input_formats_robust.py#L33

this test says we shouldn't make any API calls, and yet despite the means of actually testing whether that's the case near at hand and used in other test modules, nothing is tested - https://github.com/HzaCode/OneCite/blob/12b1dea45a2b7ddcc60b1abc6cc29984b1aefbc8/tests/test_input_formats_robust.py#L79

This whole test module tests nothing - https://github.com/HzaCode/OneCite/blob/main/tests/test_integration.py

same with this one - https://github.com/HzaCode/OneCite/blob/main/tests/test_onecite_basic.py

same with this one (those would be nice to have, and what should be tested there is whether the output formats are *correct* rather than just that an error wasn't raised) - https://github.com/HzaCode/OneCite/blob/main/tests/test_output_formats.py

this module gets close and has some real parser tests in it, but due to the LLM-generated nature of this package i actually have no confidence in whether the mocked API results are what they should be. this would be more believable using e.g. [vcr](https://pytest-vcr.readthedocs.io/en/latest/) - https://github.com/HzaCode/OneCite/blob/main/tests/test_pipeline_unit.py

since most of the tests don't test anything, that explains why my results are so bad - true to form with LLMs, I get **something,** it's just that what i get is **wrong.**
