### 2026-02-12 HzaCode/OneCite#34 @sigma-agent
Opened a focused PR: https://github.com/HzaCode/OneCite/pull/35

It tightens tests in tests/test_input_formats.py to assert parsing/identification correctness (DOI/arXiv normalization, expected metadata, and report counts), instead of only checking non-failure.
