### 2026-02-12 #267 @HzaCode
Hi @sneakers-the-rat, @giovana-morais, and @slobentanzer,

I want to express my deepest gratitude for the immense effort you’ve poured into this review.

The depth of the audit from @sneakers-the-rat and the specific functional testing from @giovana-morais is exactly the kind of rigorous feedback loop this project needed. It is a privilege to receive this level of mentorship.

I built OneCite originally to solve a specific, high-friction pain point in my own daily research workflows—handling messy citations where existing tools fell short. In my own use cases, the tool has served its purpose well. However, your review has clearly highlighted the gap between a tool that works in a specific personal context and one that is robust enough for universal community adoption.

It is clear now that while the core concept addresses a real need, the current architectural implementation requires significant refinement to account for the wide variance of input formats and edge cases expected of a public standard.

I am fully committed to closing this gap. I plan to use your feedback to refine the package's architecture, strip away logic that isn't universally robust, and significantly improve the test coverage to ensure it meets the reliability standards of pyOpenSci.

I will take the necessary time to refactor the core logic and ensure the testing is comprehensive enough to support the broader use cases required for a community tool. I will notify you here when the updated version is ready for a second look.

Thank you again for helping me elevate this project.
