### 2026-02-13 #267 @HzaCode
Hi @giovana-morais, and all,

Thank you for the continuous suggestions and the ongoing feedback loop.

Out of respect for the comprehensive nature of this review, I want to ensure I dedicate the necessary time to properly incorporate this advice and restructure the package accordingly, rather than rushing a piecemeal fix. Please expect some delay as I work to do justice to these suggestions.

I appreciate your patience.
