### 2026-02-19 6b43b3c
docs: add changelog entry for CrossRef unification (#26)
