### 2026-03-03 #267 @HzaCode
First of all, sorry everyone. I have been incredibly busy lately and haven't been able to dedicate the time needed to properly implement the structural revisions we discussed.

However, I will definitely continue to heavily refactor and improve OneCite in the background based on all your invaluable feedback. My intention for developing this tool and submitting it here was never just to get a publication, but rather to seek this exact kind of rigorous feedback and expert insight to make it genuinely better. Once the architecture and utility are robust enough to truly meet the community's standards, I hope I might have the opportunity to share it with pyOpenSci again in the future.

Even though we aren't moving forward right now, I still want to express my deepest gratitude to all of you. Thank you so much for taking the time to thoroughly audit my code and for the guidance to help me build a better tool.

Thank you again for your time and effort!
