### 2026-03-03 #267 @sneakers-the-rat
Hey! Thanks for following up. I wanted to give a bit of context from the PoV of people who are trying to maintain open peer review systems post-"AI" - mind you this is my own perspective and **not** that of pyopensci:

This right here is what I view the entire purpose of open peer review systems like pyos and joss:

> My intention for developing this tool and submitting it here was never just to get a publication, but rather to seek this exact kind of rigorous feedback and expert insight to make it genuinely better.

That's why I review and edit, to help people improve their code, to learn something myself, to provide the teaching and mentorship that I never got when I was learning to program by brute force reading code and docs by myself before even knowing that such a thing as "open source" even existed.

The challenging thing with AI written code is that it disrupts the prosocial, bidirectional help me help you process that open peer review is supposed to facilitate. Before LLMs, we wouldn't even recognize the assumption of "the author has spent more time writing this than I have spent reading it" as a foundational part of the social contract of open review, but now it's abundantly obvious. I am **not** suggesting this is what you did, but all open peer review systems are being overrun with low effort submissions where the authors spent sometimes just a few minutes of their time to generate large volumes of code and then outsource the human correction and review process to other people.

When I read code and think it can be improved, my (old) assumption is that the author has written and rewritten this, and the state of the code when I see it is at the end of some history of hard choices. The act of reviewing code is then to imagine the constraints that the author faced, to try and understand how we got here, and then reviewing and teaching is not just trying to fix the code at hand but improve the process by which it was developed so that *next time* the author can reach a little bit further. AI code disrupts this process too, because the modality of teaching is no longer the same as the modality of authorship: I can't really talk about architectural decisions, code patterns, ask about the tradeoffs the author was weighing, and so on because *there were no decisions made by the author* and *there is no means of translating feedback into future improvement.*

You seem like someone who is genuinely trying to learn and trying to improve their craft, and I appreciate that and hope that the feedback was useful. However I would recommend that you not repeat the process of vibing up a package and throwing it to other people to review after the fact. Code fundamentally a social process, and even if it wasn't your intention, the *perception* that you are trying to outsource your own labor of authorship to reviewers will be corrosive to your development and reputation. I would recommend engaging with other people early, often, seeking feedback, documenting and describing decisions you make as you make them. giving it to your friends and have them try it out first, give yourself time to work on and mature the package - *that* is the point when you should seek external, formal review: after you have demonstrably done "your share of the labor," that you are a human being seeking the expertise of other human beings. I *gladly* review code written by near-amateurs with great regularly *when i can see that they have skin in the game and have put their heart into it.*

It is unfortunate that this deliberative, thoughtful process is precisely what LLMs shortcircuit and stunt in the growth of a programmer. I imagine that the cat might be too far out of the bag to recommend not generating your code if you are learning, but lines of code appearing in a document is arguably the least important part of learning as a programmer. By all indications, I believe you are well intentioned and just want to learn, I just wanted to give you my perspective as the person on the other side giving the feedback - bluntly, I am unlikely to review another package that has been primarily generated rather than written because I don't see how it is a good use of my time or conducive to a healthy open source social ecosystem. I'm not alone in this, most nonprofit organizations that deal with review in some way are moving towards requiring evidence of human labor as a stopgap way of stemming the problems from AI. So my one recommendation is to engage socially with other programmers all along the way - people like to help other people, but people don't necessarily like to feel like they're being taken advantage of to clean up after someone else's mess, and seeing real labor done, hard problems approached, friends consulted, etc. goes a long way to dissolve that feeling.

wish you best of luck in the future
