### 2026-03-03 #267 @slobentanzer
@sneakers-the-rat Many thanks for the time you put into this review and for phrasing your conclusion so comprehensively. While I also cannot speak for pyOpenSci, I personally fully agree with the assessment and recommendations.

We do open source because we think it's the right thing to do. So far I am not willing to believe that LLMs have reached the level of self-awareness to understand this motivation (and, IMO, it will be a long way to go until some automaton can). Until they can, it is unfortunately not likely that they can autonomously operate under the same social contract as human developers.
