### 2026-03-04 #267 @HzaCode
I want to sincerely apologize again for the delayed response. I have been swamped with work recently and couldn't get back to this thread sooner.

Thank you both for taking the time to write such thoughtful and honest reflections. I read your comments carefully, and I completely agree with your perspective on the open-source social contract and the value of human-driven peer review.

*OneCite* is actually a project I have been actively developing for quite some time. Because it has proven genuinely useful in our daily workflows and shown real potential, I submitted it here hoping to get expert feedback and see where it falls short of higher architectural standards.

I deeply value and respect the hard work and strict standards you brought to my submission. As someone who also serves as a reviewer in other venues, I know firsthand the immense time, energy, and dedication it takes to conduct a thorough review.

I will take your actionable feedback to heart as I continue to develop, refine, and maintain this package for its users.
